# Fixed #11665 -- Made TestCase check deferrable constraints after each test.