… end of a test.

<https://code.djangoproject.com/ticket/11665>