@charettes 2016-02-04
I wonder if this has a significant performance impact. Some backends are really slow at checking constraints.