@jdufresne 2016-02-04
> I wonder if this has a significant performance impact. Some backends are really slow at checking constraints.

For the databases at my disposal:

* sqlite: doesn't support deferrable constraints
* PostgreSQL: constraint checking is reasonably fast, but I'm open to others' interpretations
* MySQL: doesn't support deferrable constraints

I don't have easy access to test Oracle. I am not sure how this would affect third party database backends. If anyone has insight here, I'm interested.

I can run some rough benchmarks with and without this change on those three backends and report how it affects tests performance on my box.