@charettes 2016-02-04
I know that MSSQL is notably slow at it.

If you could provide numbers for PostgreSQL it could be interesting.