@akaariai 2016-02-04
I don't believe there to be a huge performance penalty for this. Databases are excellent at checking deferred constraints - they have to be, as that is done at the end of every transaction.

Checking constraints on MSSQL is slow, but it doesn't support deferred constraints, so no problem there.

I'm a bit torn about this - on the other hand it seems we get an additional safety check, but on the other hand this will be yet another complication to test running.

Also, this has a possibility to cause problems: if you have deferrable constraints of which some are initially immediate and other initially deferred, then after the first check all of your constraints will be set to deferred. Not something to worry about I guess...

All in all, I'm mildly +0 to the approach taken here.