@jdufresne 2016-02-04
Using Python3's unittest's timing at the end of `runtests.py` I see the following with two back to back runs with a PostgreSQL database:

Without constraint checking:
Ran 10490 tests in 583.936s
Ran 10490 tests in 583.080s

With constraint checking:
Ran 10491 tests in 583.566s
Ran 10491 tests in 588.564s