@jdufresne 2016-02-04
> I'm a bit torn about this - on the other hand it seems we get an additional safety check, but on the other hand this will be yet another complication to test running.

Just to give some background on my motivation for changing this.

At some point my project had a bug in production. The bug looked as though it _should_ have been caught by tests but wasn't. I added constraint checking to a local, inherited version of `TestCase` and *boom* lots of tests started failing that would have caused developers to notice this production bug. I want to catch these sorts of bugs as early as possible next time. 

I figure the constraint is there to help us developers. If it isn't being used during tests, then we're missing out.

Are you able to elaborate on your concerns when you say "but on the other hand this will be yet another complication to test running."? Thanks.

> Also, this has a possibility to cause problems: if you have deferrable constraints of which some are initially immediate and other initially deferred, then after the first check all of your constraints will be set to deferred. Not something to worry about I guess...

Would you consider this a bug in the PostgreSQL backend's version of `check_constraints()`?