@akaariai 2016-02-10
I'm mostly worried about code complexity, and complexity of understanding what exactly the test system is doing. The added complexity in both of the cases is low, and this does have a possibility of catching real bugs, so I guess I'm +1.

The problem is that we can't write a correct version of check_constraints(). We should have something like SET CONSTRAINTS ALL DEFAULT command, but that doesn't exists. Luckily DEFERRABLE INITIALLY IMMEDIATE aren't found often.