@timgraham 2016-02-11
A similar test might be `tests/test_utils/test_transactiontestcase.py`. The current test looks a bit brittle but I guess it's difficult to make the assertion in a cross-database compatible way. Maybe it's not worth it.