@jdufresne 2016-02-11
> Maybe it's not worth it.

OK for now I have removed the brittle test. Let me know if this still requires a test and I can take another stab at it.

I believe all current feedback has been addressed. Please let me know if there is any additional feedback. Thanks.