@jdufresne 2016-02-12
> Could adding a try/finally here prevent cascading failures across the entire test suite? I think I observed the behavior when I was testing the patch earlier.

Good call, done.

I tested it with a fake "bad" test and this resolves the issue. Only the test violating the constraint fails.