@jdufresne 2016-02-13
> How about this for test_utils/test_testcase.py

Thanks Tim! I confirmed this fails before the change and passes with the change on Postgres. The commit is updated with the test.